[fix](ui) Fix Web SQL test status type
### What problem does this PR solve?

Issue Number: N/A

Related PR: #66690

Problem Summary: FE test compilation rejected a MariaDB short status flag passed to a Mockito stub whose method returns int. Widen the flag explicitly so COMPILE and Performance can pass fe-core testCompile.

### Release note

None

### Check List (For Author)

- Test: Targeted Doris Checkstyle audit and git diff check; TeamCity compile will verify fe-core testCompile
- Behavior changed: No
- Does this need documentation: No

+++ fe/fe-core/src/test/java/org/apache/doris/httpv2/websql/WebSqlStatementExecutorTest.java
@@ -271,7 +271,7 @@ public class WebSqlStatementExecutorTest {
         Mockito.when(connection.unwrap(org.mariadb.jdbc.Connection.class)).thenReturn(mariaDbConnection);
         Mockito.when(mariaDbConnection.getContext()).thenReturn(context);
         Mockito.when(context.getServerStatus()).thenReturn(
-                noBackslashEscapes ? ServerStatus.NO_BACKSLASH_ESCAPES : 0);
+                noBackslashEscapes ? (int) ServerStatus.NO_BACKSLASH_ESCAPES : 0);
     }
 
     private WebSqlLimits limits() {
